requirements.txt
tensorflow>=2.12.0
tensorflow-datasets>=4.9.2
numpy>=1.24.3
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.1
scikit-learn>=1.3.0
